+++ OldFiles/convergence_m_N.py
@@ -2,7 +2,7 @@ import numpy as np
 
 import matplotlib.pyplot as plt
 from market import Market
-from algo import lsmc
+from OldFiles.algo import lsmc
 
 
 T = 1
